Minor fixes to illustrate the workflow
Ok, so for those of you following this little tutorial - I've now looked through Mortens commits and fixed a few missing full stops and colons using github's online editor (just click on a file and it will open - README.md in this case). I can preview the modifications by clicking **preview** in the top of the editor to see what I'm actually going to commit, we don't want to introduce unnecessary errors and this is a helpful tool.

Below the editor you can then **commit** the changes you've made by giving the commit a name and a short description (this one!). Below that you choose the account/email from which you are committing, and finally you choose whether you want to commit directly into Mortens branch (remember: `mp/readme`), or if you feel the modifications are critical enough to justify a new branch based on the `mp/readme` branch. This is more useful when you modify a major branch such as the master or main development branch.