Clean up source files (#12)
* cleanup: remove executable bits

The removal was done with the following command:

  find . -type f -not -path './.git/*' -not -path './grasptest/*' -not -path './make_environment*' -perm /uog+x -exec chmod a-x {} \;

The executable flags from the make_* scripts in the root directory were
not removed, even though they are not useful either.

* cleanup: remove trailing whitespace and newlines

Modern editors generally remove these automatically, and hence it is
necessary to clean up all files. Otherwise future source change commits
will contain a lot of noise with with whitespace changes, or the
developer has to patch-add the source files, separating whitespace
changes from meaningful changes. As both are sub-optimal, the best
option is to fix all the whitespace at once.

The following commands were used to clean up the whitespace:

  find . -type f -not -path './.git/*' -not -path './grasptest/*' -exec sed --in-place -e :a -e '/^\n*$/{$d;N;};/\n$/ba' {} \;
  find . -type f -not -path './.git/*' -not -path './grasptest/*' -exec sed --in-place 's/[[:space:]]\+$//' {} \;