Minor update of environment files (#13)
* rename environment files

* Synchronized and updated/generalized the env files

The environoment files are updated with a common header, and the flags
have been updated to be slightly more generalized to fit the general
user (i.e. per and cff home folders are replaced with the $HOME env variable)

* Changed intel mpi compiler to recommended wrapper

An intel mpi comilation should be done with the wrapper mpiifort, and
not mpifort, as stated in the intel documentation [1]

[1] https://software.intel.com/en-us/mpi-developer-reference-linux-compilation-commands

* GRASP2K --> GRASP

Co-Authored-By: jongrumer <jongrumer@gmail.com>

* GRASP2K --> GRASP

Co-Authored-By: jongrumer <jongrumer@gmail.com>

* Add missing "_" in LD_LIB..

Co-Authored-By: jongrumer <jongrumer@gmail.com>

* Add missing "_" in LD_LIB... flag.

Co-Authored-By: jongrumer <jongrumer@gmail.com>

* Short installation requirements and fixed MPI_TMP

Added a short installation requirements list and modded MPI_TMP to grasp_mpi_tmp according to @mortenpi's suggestion.

* Short installation requirements and fixed MPI_TMP

Added a short installation requirement list and modified MPI_TMP according to @mortenpi's suggestion.