Added M-value and updated printing a bit
Printing for the non-default options are messy still. Working on that.

We need to agree on a form for the (non-rel) matrix value. At the moment
I decided to go for the square root of the linestrength (SA or SB) with
a phase defined by ASFA or ASFB using the SIGN() intrinsic Fortran function.

E.g., for an Ek in Coulomb gage: SIGN(DSQRT(-SA), ASFA)