Set the join in the middle of the grid when JP=0
This routine uses an algorithm that involves the orbital energy for
finding a join.  When the energy estimate is such that the criteria
for the join is not met, the program STOPS. The `STOP` has been
replaced  by  JP=NNNP/2`.  This sets the join to be the middle of
the grid.