Larger update of rmcdhf(_mpi)/setlag(mpi)
Update by Charlotte Froese Fischer, May 2019.

There was a logical error in the section determining
the Lagrange multilpliers. When two subshells are
varied simultaneously the solution is not unique so the
desired solution is set to one where the lagrange
multiplier is zero. But as soon as the orbitals
for one is fixed, a lagrange multiplier is needed.
Also, the order of the lagrange multipliers were
not in the most efficient order.

The calculation of Lagrange multipliers in the CSF
procedure for heavy elements is the most time
consuming task particularly the calculation of the
exchange term. The order of computation has been
modified so that POTL and XCH functions are called
the minimum number of times.  In Francine (Fr)
45 calls to exchange were reduced to 30 call. When
correlation is included, the XCH function includes
contributions from all the interactions as well as
exchange.