Update outer boundary check with J .EQ. NNNP-1
Update by Charlotte Froese Fischer, May 2019

When the algorithm for finding the outward boundary
condition did not find an end-point, because of poor initial
estimates, an error message claimed the range was insufficient.
In the present code it is defined to be NNNP-1.  An
appropriate "end point" is determined in subsequent
iterations.  The range NNNP=590 has been found adequate
for all test cases.