Patch July 2019 (#19)
##  CFF Patch: Improving RMCDHF and extending terms to J = 17/2 (l, m-)

In May @cffischer sent me a collection of updates, fixes and general improvements that I've put together into a single patch. The most important fixes are related to how the Lagrange multipliers are calculated in RMCDHF, and she has also extended the allowed terms on step to `J=17/2 (l, m-)`. I've also added some minor fixes to the rwfnmchfmcdf code and the ldlbl1 and 2 files in the transition codes which should remove some warnings during compilation.

Check the patch out and see what you think, review the code and let's get back to this in a week or so for merging. 

A description of the mods by Charlotte follows.
/Jon

### Patch summary (CFF)
'RMCDHF' is not robust in that accurate initial estimates are needed.
When algorithms fail, the messages are misleading, and in some instances 
the program simply stops. Several changes have been introduced that 
increase the robustness and the efficiency.

### Tests
The changes have been tested for a correlation study for Li 1s(2)2s(1) J=1/2 for wavefunctions from SDT excitations to nl=9l. 

### Libmod - increase J of allowed terms.
  
  1) `terms_C.f90`

Changes have been introduced that increase the maximum `J` to 17/2 
or `l` or `m-`. 
  
### Lib9290 - joining point evaluation has been improved

  1)  `setpot.f90`
 
This routine uses an algorithm that involves the orbital energy for
finding a join.  When the energy estimate is such that the criteria
for the join is not met, the program STOPS. The `STOP` has been 
replaced  by  `JP=NNNP/2`.  This sets the join to be the middle of the grid.

  2)  `start.90`

Remove the comment about accuracy because the problem is not related to
accuracy and is misleading.  

### RMCDHF

  1) `setlag.f90` and `setlagmpi.f90`

There was a logical  error in the section determining
Lagrange multilpliers.  When two subshells are
both varied the solution is not unique so the
desired solution is one where the lagrange
multiplier is zero.  But as soon as the orbitals
for one is fixed, a lagrange multiplier is needed.
Also, the order of the lagrange multipliers were 
not in the most efficient order. 

The calculation of Lagrange multipliers in the CSF 
procedure for heavy elements is the most time 
consuming task particularly the calculation of the 
exchange term. The order of computation has been 
modified so that POTL and XCH functions are called 
the minimum number of times.  In Francine (Fr) 
45 calls to exchange were reduced to 30 call.`When 
Correlation is included, XCH function includes 
contributions from all the interactions as well as 
exchange.

  2) `in.f90`     

When the algorithm for finding the outward boundary
condition did not find an end-point, because of poor initial
estimates an error message claimed the range was insufficient.
In the present code it is defined to be `NNNP-1`.  An
appropriate "end point" is determined in subsequent
iterations.  The range `NNNP=590` has been found adequate
for all test cases.

### Tools - increase dimension of the tools codes

Many of the routines in the `tool` library, do not use the GRASP 
parameters for determining dimensions and are subject to error.  In 
particular, "lines" of input data have character dimension of 100.  For 
heavy elements, this may not be sufficient and the program does not 
detect that information is lost.

The two corrected cases are,
   1) `rmixaccumulate.f90`
   2) `rmixextract.f90`

but there are probably more.

Actually, it is not a correction really since this just changes `100 => 200`. But 
there are several other cases that probably also are wrong.  Unless 
someone else updates the online GRASP2018 code, Charlotte said she will do that as soon, possibly this summer. `rmixextract.f90` also had another error that she fixed.

## Additional fixes (JG)
  1) `tools/rwfnmchfmcdf.f90`

Bugfix: initialization loop of arrays were done in an old fashioned style, but also from element 0 which gave a warning during compilation. This is now changed to simple `A = 0` statements.

  2) `rtransition(_mpi)/ldlbl1(2).f90`

Fixed legacy-style read statements.