Add support for CFF's new short csl format
CFF: I proposed a shorter and much more informative format
for the <name>.c file. Basically, the structure is by configuation,
but for each configuration all the couplings are shown along with
critical seniority numbers. So far, our code does not support CSFs where
extra information is needed. This will be important for the g-shell
which I think someone should work on with Gediminas' new rangular.

If GRASP were a well designed program, changing the format for reading
<name>.c would require a change of only one "lod" routine, but GRASP is
not well designed so little changes were needed in quite a few places.