Slightly simplify the Makefile system (#30)
* Create a root Makefile that builds everything

Each library and program still has their own separate Makefile which
depends on variables set by the root Makefile.

The environment scripts are no longer necessary -- the variables can be
overriden in Make.user. However, it should be noted that if variables
like FC are set as an environment variable, they will still override the
defaults.