Improve handling of MPI_TMP in cpath (#37)
* Fix handling of MPI_TMP in cpath

cpath has an arbitrary limit of 64 characters for the length of the
MPI_TMP environment variable. What is worse, if the variable is longer
than that, cpath will fall back to the default /scratch/$USER value,
instead of throwing an error.

With this patch, the original arbitrary limit is remove and the helper
function getenv_allocating will just allocate the necessary amount of
memory to store the value of the variable.

There is still an arbitrary limit -- 128 characters, imposed by via the
lendisk0 variable to the disk variable. disk is the variable used to
propagate the value to slave nodes, and fixing the length handling there
is beyond the scope of this patch.

* Add simple integration test for MPI_TMP handling

This has two functions: (1) it tests the MPI_TMP handling (specifically
tests whether long paths work), and (2) it functions as a simple
integration test for the basic binaries.

Note that currently the mpi-longpath fails due to a bug in cpath().

* Update comment in integration test script

* Make sure that MPI_TMP is to _too_ long either

We have another 128 arbitrary limit, which we may accidentally hit if we
put $MPI_TMP under $PWD.