New version of RCSFINTERACT (2021)
Critical bug fix by G. Gaigalas (May, 2021)

This version addresses a bug related to mismatching peel sheel ordering
in the multireference and full list. This can happen for larger, more
complex systems where correlation orbitals added in higher layers, might
end up before spectroscopic orbitals in the list of peel shells, using
standard ordering. This can lead to that a lot of CSF's are removed in
the reduction that should in fact be in there, just that the peel order
is different. The present proposed fix by Gediminas adds an
additional check for this which increases the runtime, but is necessary
to avoid critical mistakes that may pass unnoticed.

This bug fix is to consider as critical and should be adopted by all
users who make use of CSF list reductions using this code immediately.