New version of RCSFINTERACT (2021) (#65)
* New version of RCSFINTERACT (2021)

Critical bug fix by Gediminas Gaigalas <Gediminas.Gaigalas@tfai.vu.lt> (May, 2021)

This version addresses a bug related to mismatching peel shell ordering
in the multireference and full list. This can happen for larger, more
complex systems where correlation orbitals added in higher layers, might
end up before spectroscopic orbitals in the list of peel shells, using
standard ordering. This can lead to a lot of CSF's being removed in
the reduction that should in fact be in there, due to the fact that the peel order
is different, i.e. mismatching coupling order. The present fix adds an
additional check for this which increases the runtime but is necessary
to avoid critical mistakes that may pass unnoticed.

This bug fix is to consider critical and should be adopted by all
users who make use of CSF list reductions using this code immediately.