rwfntotxt + rwfnpyplot  -- new plotting tools (Grumer) (#82)
* New plotting tools

For improved flexibility when it comes to plotting GRASP orbtial wavefunctions, I suggest we adopt the same format as used by DBSR_HF where the orbitals are dumped as columns in a text file. This file can
then be used by a plotting tool of choice. Hope they will be of use to you all - it is indeed important to check your orbitals by eye!

New tools:
- rwfntotxt: converts a GRASP binary orbital file to ASCII
  Input: rwfn.inp --> Output: rwfn.plot
- rwfnpyplot: a python script that can be used to conveniently plot orbitals from a .plot file (from either DBSR_HF or GRASP/rwfntotxt)
  Run with command line arguments - for more info, just run
  rwfnpyplot --help