rmixextract -- new extraction mode (Gaigalas) (#84)
* New extraction mode for rmixextract - by G. Gaigalas

This code extracts CSF's if a mixing coefficient is larger than a given cut-off value.

This was previously done for across all eigenstates together - i.e. if the mixing coeff. for a given CSF was larger than the cut-off in any of the eigenstates, it was extracted. Now, a second option is added where this can be done also for each eigenstate separately. This facilitates e.g. extractions of CSF's for more problematic eigenstates, i.e. you can weight your calculation to include more important interactions for the more challenging eigenstates.

The code now outputs both a rcsf.out file (containing the total extracted list) and an .rmx file which lists coefficients and CSF's for each individual eigenstate.

Co-authored-by: Gediminas Gaigalas <Gediminas.Gaigalas@tfai.vu.lt>